gather_data/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
gevent==23.9.1
greenlet==3.0.3
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
websocket-client==1.7.0
zope.event==5.0
zope.interface==6.1
